yet another small refactoring
