+++ lib/features/settings/data/model/implementations/delete_confirmation_setting.dart
@@ -1,4 +1,4 @@
-part of '../../../../../core/services/settings/service/settings_service.dart';
+part of '../../../../../core/services/settings/settings_service.dart';
 
 class DeleteConfirmationSetting implements Setting<bool> {
   final SettingsStorage _storage;
