+++ lib/features/settings/domain/state_management/locale/locale_cubit.dart
@@ -1,6 +1,6 @@
 import 'package:bloc/bloc.dart';
 import 'package:get_it/get_it.dart';
-import 'package:todo_list/core/services/settings/service/settings_service.dart';
+import 'package:todo_list/core/services/settings/settings_service.dart';
 
 class LocaleCubit extends Cubit<String> {
   final _locale = GetIt.I<SettingsService>().locale;
