+++ lib/features/settings/domain/state_management/theme/theme_cubit.dart
@@ -1,7 +1,7 @@
 import 'package:bloc/bloc.dart';
 import 'package:flutter/material.dart';
 import 'package:get_it/get_it.dart';
-import 'package:todo_list/core/services/settings/service/settings_service.dart';
+import 'package:todo_list/core/services/settings/settings_service.dart';
 
 class ThemeCubit extends Cubit<ThemeMode> {
   final _theme = GetIt.I<SettingsService>().theme;
