+++ lib/features/todo/data/repositories/todo_repository_impl.dart
@@ -6,7 +6,7 @@ import 'package:todo_list/features/todo/data/sources/remote/remote_source/remote
 import 'package:todo_list/features/todo/domain/entities/todo_entity.dart';
 import 'package:todo_list/features/todo/domain/repository/todo_repository.dart';
 import 'package:get_it/get_it.dart';
-import 'package:todo_list/core/services/settings/service/settings_service.dart';
+import 'package:todo_list/core/services/settings/settings_service.dart';
 
 part '../mapping_extensions/entity_to_dtos.dart';
 part '../mapping_extensions/local_to_entity.dart';
